+++ cuda/src/main/java/org/bytedeco/javacpp/presets/cublas.java
@@ -34,7 +34,8 @@ import org.bytedeco.javacpp.tools.InfoMapper;
  * @author Samuel Audet
  */
 @Properties(inherit = cuda.class, value = {
-    @Platform(define = {"CUBLASAPI", "CUBLAS_V2_H_"}, include = {"<cublas_api.h>", "<cublas.h>", "<cublasXt.h>"}, link = "cublas@.9.2")},
+    @Platform(define = {"CUBLASAPI", "CUBLAS_V2_H_"}, include = {"<cublas_api.h>", "<cublas.h>", "<cublasXt.h>"}, link = "cublas@.9.2"),
+    @Platform(value = "windows-x86_64", preload = "cublas64_92")},
         target = "org.bytedeco.javacpp.cublas")
 @NoException
 public class cublas implements InfoMapper {