+++ cuda/src/main/java/org/bytedeco/javacpp/presets/cufftw.java
@@ -34,7 +34,8 @@ import org.bytedeco.javacpp.tools.InfoMapper;
  * @author Samuel Audet
  */
 @Properties(inherit = cufft.class, value = {
-    @Platform(include = "<cufftw.h>", link = "cufftw@.9.2")},
+    @Platform(include = "<cufftw.h>", link = "cufftw@.9.2"),
+    @Platform(value = "windows-x86_64", preload = "cufftw64_92")},
         target = "org.bytedeco.javacpp.cufftw")
 @NoException
 public class cufftw implements InfoMapper {